moved auth_service to item_spec
and it works for route planner
